Extract navigation related things into own class
